requirements.txt
discord.py
openai
python-dotenv
aiohttp
flask
sqlalchemy
psycopg2-binary
stripe>=5,<6
